apps/ingestion/requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.9
minio==7.2.3
prometheus-client==0.19.0
pyyaml==6.0.1
